refactor: extract common persona service
